pretty_print_stacks: prevent invalid address arguments
If a return address is zero, the script will pass -1 to addr2line which
it considers an invalid flag.
Zero is an unlikely return address, but we might see it, for example
because the backtrace is unreliable on s390x if there is an interrupt
frame on the stack.

This is an addition to test rebase hooks. [Pavan]

Fixes: a9143a24 ("scripts: pretty print stack traces")
Signed-off-by: Nina Schoetterl-Glausch <nsg@linux.ibm.com>
Tested-by: Thomas Huth <thuth@redhat.com>
Message-Id: <20230404185048.2824384-2-nsg@linux.ibm.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>