x86 AMD SEV-SNP: Setup a new page table and install level 1 PTEs
Currently, there is no support for page allocator in the guest
supporting UEFI. Setup a new page table and ensure the GHCB page is
unencrypted, i.e., its c-bit should be unset, before the page is ready
for its allocation. Besides, KVM Unit Tests installs 2M pages, i.e.,
only level 2 PTEs are provided. Unsetting the c-bit of a level 2 PTE
would be detrimental to the guest as it will read a garbage PTE. So the
solution to this problem is to install 4K PTEs and do any page state
change operations on level-1 PTEs, that are to follow in the next set of
patches.

Reported-by: Kim Phillips <kim.phillips@amd.com>
Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>