x86 AMD SEV: Include AMD SEV tests into x86/unittests.cfg
As of today, AMD SEV tests are not part of run-tests. This patch
enables the support for these tests as part of run-tests which get
PASSed under UEFI and get SKIPped appropriately under Non-UEFI.

A caveat here is that AMD SEV tests need reconfiguration before and
after they are run, the support of which is also provided in this patch.
This serves as a temporary fix but a permanent solution to this is still
required.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>