x86 AMD SEV-SNP: Convert guest pages from private to shared using GHCB MSR protocol
As mentioned in the GHCB spec (Section 2.3.1 GHCB MSR protocol), the
SEV-SNP guest VM issues 4K page state changes request to hypervisor to
convert KUT's newly allocated pages to shared pages using GHCB MSR protcol.

The purpose of this test is to determine the state of the pages is being
changed to shared by the hypervisor by reading/writing data to the pages
with the c-bit cleared.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>