x86 AMD SEV-SNP: Convert a guest page from private to shared using GHCB MSR protocol
The SEV-SNP guest VM issues a 4K page state change request to hypervisor
to convert a KUT's newly allocated page to a shared page using GHCB MSR
protcol.

The purpose of this test is to determine if the c-bit of the page trying
to be modified is unset properly and the state of the page is being
changed as expected by the hypervisor.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>