x86 AMD SEV-SNP: Convert a newly allocated page from private to shared
The SEV-SNP guest VM issues a page state change request to hypervisor to
convert a KUT's newly allocated page to a shared page.

The purpose of this test is to determine if the c-bit of the page trying
to be modified is unset properly and the state of the page is being
changed as expected by the hypervisor.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>