x86 AMD SEV-SNP: Page state change operation from shared to private
The SEV-SNP guest VM issues a page state change request to hypervisor to
convert a hypervisor owned shared page back to private. Guest then
issues a 'pvalidate' instruction to validate the page after its
conversion.

The purpose of this test is to determine if the state of the page is
being changed as expected by the hypervisor and c-bit is set properly
for the page that is now private.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>