x86 AMD SEV-SNP: Introduce AP bringup support using AP creation NAE
AMD SEV-SNP guests require AP creation NAE event to cause the
creation/destruction of, or a change to, the register state of an AP,
which serves as a method for booting an AP.

This patch brings SEV-SNP APs online where APs enter into a long mode.
KVM unit tests use pre_boot_apic_id() while setting up the TSS for AP
within GDT. This raises #VC exception and therefore #VC handler needs to
be setup for the APs.

The following patches in these series introduce support for #VC handler
for APs.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>