x86 AMD SEV-SNP: Introduce AP bringup support using AP creation NAE
AMD SEV-SNP guests require AP creation NAE event to cause the
creation/destruction of, or a change to, the register state of an AP,
which serves as a method for booting an AP.

This patch brings SEV-SNP APs online where APs enter into a long mode.

After this patch is applied, kvm-unit-tests under UEFI with smp > 1 for
SEV-SNP guest fails after BSP enters cpu_relax() (in lib/x86/smp.c).
This is because setup_gdt_tss() uses pre_boot_apic_id() while setting up
the TSS for AP within GDT. This raises #VC exception for the AP, the
support for which will be introduced in the following patches.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>