x86 AMD SEV-SNP: Convert a page from private to shared
The SEV-SNP guest VM's AP issues a page state change request to
hypervisor to convert the newly allocated private GHCB page to a shared
one.

This also serves as a test case to verify a page state change from
private->shared. Add a check in x86/amd_sev.c as a test case to verify
the page state change.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>