x86 AMD SEV-SNP: Perform per-CPU GHCB GPA registration
Each vCPU needs to have its GHCB page registered for it to be able to
talk to the hypervisor.

With this patch, all the APs should be able to come up online and should
be in the HLT state thereafter waiting for respective IPIs. Thereafter,
BSP should be able to come out of cpu_relax() (as total_cpu_count
equals number of online CPUs) and run the efi_main() to complete the
remainder of KUT tests.

This test has been carried out with # of APs ranging from 2-10 and all
the APs were able to come up online and get into HLT (ap_online()).

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>