x86 AMD SEV-SNP: Change guest pages from private to shared using GHCB NAE
Perform page state change conversions on a user inputted number of pages
from private to hypervisor-owned. The page state change NAE event allows
for SEV-SNP guest to request page state changes to hypervisor using GHCB
protocol.

The test introduces support for both 4K as well as 2M large page,
depending on the order (1 << order) specified by the user and whether
the address of the page is 2M aligned or not. If 2M range is backed by a
4K page, pvalidate fails with FAIL_SIZE_MISMATCH error. In such a case,
the guest tries to pvalidate all the 4K entries in this 2M range.

Write data to the encrypted pages before private->shared conversion,
performs the page state conversions, unsets the c-bits on these pages,
and then writes data to the un-encrypted pages after private->shared
conversions to confirm we can write data to the guest pages before and
after page state changes.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>