x86 AMD SEV-SNP: Test intermix to shared PSCs on 2M shared page
The test performs the following actions:
1. Allocates a 2M private page (512 4K entries) and convert the range to
   a 2M shared range.
2. Creates an intermix of shared and private entries within 2M range by
   converting 256 entries to private and leaving the rest as shared.
3. Converts from intermixed state to 2M private.
4. Converts the 2M range from private to shared.

The guest issues an explicit page state change to private on
a 2M range from the intermixed state (step-3). Without this explicit
conversion to private, guest will trigger an RMP fault for implicit
page state changes on all the entries that are currently in shared state
in the intermix, and converts them to private implicitly, before
converting the 2M range to shared from intermixed state.

Suggested-by: Michael Roth <michael.roth@amd.com>
Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>