x86 AMD SEV-SNP: Perform PSCs from shared to private using GHCB MSR protocol
The SEV-SNP guest VM issues page state changes request to hypervisor to
convert hypervisor-owned 4K shared pages back to private (guest-owned)
using GHCB MSR protocol. Guest then issues a 'pvalidate' instruction to
validate the pages after the conversions.

The purpose of this test is to determine the state of the pages is being
changed to private by the hypervisor by reading/writing data to the
pages with C-bit set before conversions.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>