Rectify re-validation logic and test on 2M Shared Page
Any attempt to issue a 'pvalidate' instruction on a shared page (2M/4K),
in this case, a 2M page is leading to error below:

qemu-system-x86_64: warning: memory fault: GPA 0x3bc00000 size 0x1000 flags 0x8
FAIL: Expected state: SHared

My understanding: Once if a page is turned to shared from private, any
attempt to issue a re-validation to it to rescind the validated bit
again should only result in a PVALIDATE_FAIL_NOUPDATE. And if in case,
it is a large-page (2M), it could be possible that RMP entry is a 4K
one, in which case we hit PVALIDATE_FAIL_MISMATCH, which is when the
guest issues a 'pvalidate' on 4K entries of this 2M large page.

The c-bit unsetting becomes important only when the guest is attempting
to write something into this page after it is a shared page, which
should not matter here as we are attempting a re-validate immediately
after the 2M page has been converted from private to shared.