Add re-validation logic and test on a 4K private and shared page
Re-issue 'pvalidate' on a private 4K page to determine if it hits
PVALDIATE_FAIL_NOUPDATE. PASS the test in that case.

Then, once the page has been converted to shared using GHCB MSR,
Re-issue 'pvalidate' on now-shared 4K page to ensure
PVALIDATE_FAIL_NOUPDATE gets hit again. However, this fails to happen as
the following error pops up:

qemu-system-x86_64: warning: memory fault: GPA 0x163c000 size 0x1000 flags 0x8
PASS: Expected State: Shared

The output of the sequence described above is posted below:

Private->Shared conversion test.
PASS: Expected State: Private
PASS: Write to encrypted page before private->shared conversion
qemu-system-x86_64: warning: memory fault: GPA 0x163c000 size 0x1000 flags 0x8
PASS: Expected State: Shared
(HANGS....)

My understanding: Once if a page is turned to shared from private, any
attempt to issue a re-validation to it to rescind the validated bit
again should only result in a PVALIDATE_FAIL_NOUPDATE. And if in case,
it is a large-page (2M), it could be possible that RMP entry is a 4K
one, in which case we hit PVALIDATE_FAIL_MISMATCH, which is when the
guest issues a 'pvalidate' on 4K entries of this 2M large page.

The c-bit unsetting becomes important only when the guest is attempting
to write something into this page after it is a shared page, which
should not matter here as we are attempting a re-validate immediately
after the page has been converted from private to shared.