x86 AMD SEV-SNP: Enable SEV-SNP support
Incorporate support for SEV-SNP enablement. Provide a simple activation
test to determine whether SEV-SNP is enabled or not.

SKIP this activation test if the guest is not an SEV-SNP guest.

Besides, for SEV-SNP, the requirement is that SEV-ES and SEV be enabled.
In addition, setup_vc_handler() is common to both SEV-ES and SEV-SNP.
Therefore, call setup_vc_handler() only when SEV-ES is enabled.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>