x86 AMD SEV-ES: Set GHCB page attributes for a new page table
SEV-ES/SNP guest uses GHCB page to communicate with the host. Such a
page should remain unencrypted (its C-bit should be unset in the guest page
table). Therefore, call setup_ghcb_pte() in the path of setup_vm() to ensure
C-bit of GHCB's pte is unset, for a new page table that will be setup as
a part of page allocation for UEFI-based SEV-ES/SNP tests later on.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>