x86 AMD SEV-SNP: Change guest pages from Shared->Private using GHCB NAE
Perform page state conversions to private on a range of pages that are
already in a hypervisor-owned state.

The test introduces support for both 4K pages as well as 2M large pages
depending on the order specified.

Perform a write operation on the pages while they are shared pages. After
conversion, run a re-validation test on one of the converted pages to
ensure the page state is private.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>