x86 AMD SEV-SNP: Test-1: Perform Intermix to 2M Private PSCs
The test performs the following actions:
1. Allocates a 2M private page (512 4K entries) and converts the entire
   range to shared.
2. Performs a write operation on these un-encrypted pages.
3. Performs partial page state change conversions on first 256 4K
   entries and conducts a re-validation test on one of these now-private
   entries to determine whether the current page state is private or
   not.
4. Also conducts a write test on the 256 4K shared entries (with C-bit
   unset on PMD) to ensure these are in shared state.
5. Converts the whole 2M range from an intermixed state to private and
   runs a re-validation check on the now-private 2M page.

Since the test performs partial page state changes within 2M range,
pvalidate would result in a failure in the form of
PVALIDATE_FAIL_NOUPDATE since the test would perform same page state
change operation on the entries that are already in the same state. Do
not treat RMP entry validated bit unchanged as an error for this test.

The primary goal of this test is to determine whether a shared->private
conversion on a 2M range containing a mix of shared and private entries
is handled properly by the hypervisor or not. Such a PSC from an
intermixed state may never take place in a conventional SEV-SNP guest.

Suggested-by: Michael Roth <michael.roth@amd.com>
Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>