x86 AMD SEV-SNP: Test Private->Shared page state changes using GHCB MSR
As mentioned in the GHCB spec (Section 2.3.1 GHCB MSR protocol), the
page state change GHCB MSR protocol is used to convert a 4K page from
private to shared or vice-versa. Add support for this test by allocating
a total of 1024 4K pages to ensure the test can handle large cases.

The purpose of this test is to determine whether the hypervisor changes
the page state to shared when using the MSR protocol. Before the
conversion test, ensure the state of the pages are in an expected state
(i.e., private) by issuing a re-validation on one of the newly allocated
page to determine the expected state of the page matches with the page's
current state. Report failure if the expected page state is not private.

After the page state conversion to shared by the hypervisor, ensure the
state of the pages are in shared by writing data to these pages while
the C-bit in its PTEs is not set. Report a failure otherwise.

Provide support for cleaning up the physical pages by converting the
pages to default guest-owned state before freeing them.

Import GHCB MSR PSC related and pvalidate definitions from upstream
linux. (arch/x86/inlcude/asm/sev-common.h and
arch/x86/include/asm/sev.h)

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>