x86 AMD SEV-SNP: Change guest pages from shared to private using GHCB NAE
Perform page state change conversions to private on a range of pages
that are already in hypervisor-owned state.

The test introduces support for both 4K pages as well as 2M large pages
depending on the order and the page size specified.

While at it, make changes to pvalidate_pages() to not treat
PVALIDATE_FAIL_NOUPDATE as an error when converting the already private
pages as a part of cleanup process.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>