x86 AMD SEV-SNP: Change guest pages from Intermix->Private using GHCB NAE
The tests perform the following actions:
1. Allocates a 2M private page (512 4K entries) and converts the entire
range to shared.
2. Performs a write operation on these un-encrypted pages.
3. Performs partial page state change conversions on the first 256 4K
entries and conducts a re-validation test on one of these now-private
entries to determine whether the current page state is private or not.
4. Converts the whole 2M range from an intermixed state to private and
perform a re-validation check on the now-private 2M page.

The goal of this test is to ensure 2M page state changes are handled
properly even if 2M range contains a mix of private/shared pages.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>