x86 AMD SEV-SNP: Change guest pages from Shared->Private using GHCB NAE
Convert the same pages back to private that were converted to shared.

The test handles both 4K and 2M large pages depending on the order and
the page size specified.

While at it, make changes to pvalidate_pages() to not treat
PVALIDATE_FAIL_NOUPDATE as an error when converting the already private
pages as a part of cleanup process.

Signed-off-by: Pavan Kumar Paluri <papaluri@amd.com>